Homework 1 CMPSC 132 PROG & COMP II
In this homework, I implemented a suite of Python functions to tackle a variety of problems. I created a rectangle function to determine one of the rectangle's dimensions based on its perimeter and area, an invert function to swap keys and values in a dictionary while handling duplicates, and a successors function to build a successor (Markov chain) dictionary from a text file. Additionally, I developed a getPosition function to find the position of a digit in a number, a hailstone function to generate the hailstone sequence for a given number, and a largeFactor function to compute the largest proper factor of a number. This assignment reinforced my skills in using control structures, functions, file handling, and data manipulation in Python.