Homework 4 CMPSC 132 PROG & COMP II
In this homework, I built an anagram finder using a Binary Search Tree (BST). I modified the BST's insert method so that each node stores a key generated by sorting the letters of a word, and a list of words that match that key. The tree is then used to search for anagrams by comparing these sorted keys, returning all matching words if they exist. Additionally, I implemented an in-order traversal to display the tree's content, reinforcing my understanding of BSTs and string manipulation in Python.