LAB 3 CMPSC 132 PROG & COMP II
In Lab 3, I implemented several recursive functions without using any loops or global variables. I created a function to recursively count the occurrences of an item in a list, another to recursively replace all instances of a value with a new one, a function to recursively process a list by cutting it based on negative values, and one to recursively remove repeated neighboring digits from a number. This assignment deepened my understanding of recursion and functional programming techniques in Python.