Homework 5 CECS 3200 ASSEMBLY LANGUAGE PROGRAMMING
In this homework, I developed a menu-driven calculator program in Assembly language. The program displays a menu with options for addition, subtraction, division, and multiplication, then prompts the user for two numbers. Using macros for streamlined code—such as for input handling and arithmetic operations—it performs the chosen calculation and displays the result via DOS interrupts. This project reinforced my understanding of low-level programming concepts, including memory models, data manipulation, and the use of assembly macros.