Lab 04 CECS 2223 COMPUTER PROGRAMMING II LAB
In Lab 4, I developed a Sodas class that encapsulates soda attributes such as name and price while tracking the number of instances with a static variable. I overloaded the greater-than-operator to compare soda prices. Then, I wrote helper functions to populate an array of Sodas objects based on user input, display the objects in a formatted table, and sort the array using a bubble sort algorithm. This lab reinforced my skills in object-oriented programming, operator overloading, and array manipulation in C++.